Python_Server/requirements.txt
flask
firebase-admin
python-dotenv
pydantic
Pillow
opencv-python
numpy
google-generativeai